[Oyl3D] Housekeeping

+++ Source/Oyl3D/Core/Core/Application/CommandLine.cpp
@@ -173,7 +173,7 @@ namespace Oyl
 			std::string name = arg.substr(namePos, equalsPos - namePos);
 
 			// Peek the next argument, if it is a value, use it for this argument
-			size_t nextIndex = i + 1;
+			int nextIndex = i + 1;
 			if (value.empty() && nextIndex < a_argc && !isArgument(a_argv[nextIndex]))
 			{
 				value = a_argv[++i];
